+++ server/apps/payroll/use_cases/work_breaks/delete.py
@@ -2,15 +2,26 @@ from dataclasses import dataclass
 
 from rest_framework import serializers
 
+from apps.core.application.errors import AccessDeniedApplicationError
 from apps.core.application.use_cases import BasePresenter, BaseUseCase
 from apps.payroll.models.work_break import WorkBreak
+from apps.payroll.services.work_break.allowed import can_manage_work_break
+from apps.users.models import User
+
+
+@dataclass(frozen=True)
+class WorkBreakDeleteData:
+    """Delete work break data."""
+
+    work_break: int
 
 
 @dataclass(frozen=True)
 class InputDto:
     """Delete work break input dto."""
 
-    work_break: int
+    data: WorkBreakDeleteData  # noqa: WPS110
+    user: User
 
 
 class InputDtoSerializer(serializers.Serializer):
@@ -31,9 +42,12 @@ class UseCase(BaseUseCase):
     def execute(self, input_dto: InputDto) -> None:
         """Main logic here."""
         validated_data = self.validate_input(
-            input_dto,
+            input_dto.data,
             InputDtoSerializer,
         )
 
         work_break = validated_data["work_break"]
+        if not can_manage_work_break(work_break, input_dto.user):
+            raise AccessDeniedApplicationError()
+
         work_break.delete()
